Updating to v0.2
Lots of new features, bug fixes, code improvements and new textures.
A capsule can now rotate at y = 0. It has a new attribute, landingPoint, that contains the y level at which the capsule will have reached the bottom if it isn't moved; this attribute is used to display a shadow.