Updating to v0.2
Lots of new features, bug fixes, code improvements and new textures.
The canvas now draws menus (start menu, pause menu and "you lost" menu) with buttons. There are informations at the top of the screen. The canvas needed to be adapted to the new texture system. A lot of stuff has been simplified with functions and variables.